6-med-fine-tuning-llm/pyproject.toml
[project]
name = "med-fine-tuning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torchao==0.16.0",
    "fbgemm-gpu-genai==1.4.2",
    "transformers==4.55.4",
    "trl==0.22.2",
    "datasets==4.3.0",
    "unsloth-zoo==2025.11.6",
    "unsloth==2025.11.6",
    "modelscope==1.33.0",
    "addict>=2.4.0",
    "matplotlib>=3.10.8",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true